finish graph for active users
