+++ src/main/java/Client/Admin/Views/Components/ActiveUsersLineChart.java
@@ -1,0 +1,53 @@
+package Client.Admin.Views.Components;
+
+import Client.Admin.Repository.SessionRepository;
+
+import java.util.*;
+
+public class ActiveUsersLineChart extends BaseLineChart {
+
+    public ActiveUsersLineChart(String title, String xLabel, String yLabel) {
+        super(title, xLabel, yLabel);
+    }
+    @Override
+    protected void updateDataPoints() {
+        dataPoints = new ArrayList<>();
+
+        // Get the sessions for the selected year
+        List<Map<String, Object>> sessions = sessionRepository.getSessionsForYear(year);
+
+        // Create a map to store the count of active users for each month
+        Map<Integer, Set<String>> activeUsersPerMonth = new HashMap<>();
+
+        // Iterate over each session
+        for (Map<String, Object> session : sessions) {
+            Date loginTime = (Date) session.get("login_time");
+            Date logoutTime = (Date) session.get("logout_time");
+
+            Calendar cal = Calendar.getInstance();
+            cal.setTime(loginTime);
+
+            // Iterate for each month between login_time and logout_time (inclusive)
+            while (cal.getTime().before(logoutTime) || cal.getTime().equals(logoutTime)) {
+                int month = cal.get(Calendar.MONTH) + 1; // Calendar months are 0-based
+
+                // Get or create the set for the month
+                Set<String> usersForMonth = activeUsersPerMonth.computeIfAbsent(month, k -> new HashSet<>());
+
+                // Add the user to the set (only if not already counted for this month)
+                usersForMonth.add((String) session.get("username"));
+
+                cal.add(Calendar.MONTH, 1); // Move to the next month
+            }
+
+        }
+
+        // Convert the map entries to a list of entries for plotting
+        for (Map.Entry<Integer, Set<String>> entry : activeUsersPerMonth.entrySet()) {
+            dataPoints.add(new AbstractMap.SimpleEntry<>(entry.getKey(), entry.getValue().size()));
+        }
+    }
+
+
+
+}
